2.1.0 (1517) Ready for App Store
Changes:
* Fixed issue where preferences and about windows would not come to
front if open and in the background.
* Fixed a bug where AppleScript command active did not work.
