+++ RocketFuel/AppDelegate.swift
@@ -15,16 +15,7 @@ class AppDelegate: NSObject, NSApplicationDelegate {
   
   func applicationDidFinishLaunching(aNotification: NSNotification) {
     Preferences.registerDefaults()
-    // Load hot key
-    if let hotKeyDict = Preferences.value(forKey: PreferencesType.ActivationHotKey) as? NSDictionary {
-      if let keyCode = hotKeyDict.valueForKey("keyCode") as? Int {
-        let modifierFlags = hotKeyDict.valueForKey("modifierFlags") as? Int ?? 0
-        let readable = hotKeyDict.valueForKey("readable") as? String ?? ""
-        print(modifierFlags)
-        self.activationHotKey = HotKey(keyCode: keyCode, modifier: modifierFlags, readable: readable, action: self.applicationShouldChangeState)
-        HotKeyCenter.sharedManager.register(self.activationHotKey)
-      }
-    }
+    self.applicationShouldLoadHotKey()
   }
   
   func applicationShouldChangeState() {
@@ -43,21 +34,6 @@ class AppDelegate: NSObject, NSApplicationDelegate {
     // This function is called either from the Preferences Window or an apple script.
     Preferences.save(level, forKey: PreferencesType.StopAtBatteryLevel)
     self.statusItemController.shouldDeactivateOnBatteryLevel = level
-  }
-  
-  func applicationShouldRegisterHotKey(hotKey: HotKey) {
-    HotKeyCenter.sharedManager.unregister(self.activationHotKey)
-    
-    if hotKey.keyCode > -1 {
-      hotKey.action = self.applicationShouldChangeState
-      HotKeyCenter.sharedManager.register(hotKey)
-
-      let dict = NSDictionary(dictionary: ["keyCode": NSNumber(integer: hotKey.keyCode), "modifierFlags": NSNumber(integer: hotKey.modifier), "readable": hotKey.readable])
-      
-      Preferences.save(dict, forKey: PreferencesType.ActivationHotKey)
-    } else {
-      Preferences.remove(PreferencesType.ActivationHotKey)
-    }
   }
   
 }
